Restructure redux files
