+++ web/src/Application.jsx
@@ -45,12 +45,12 @@ import SessionWebsocket from './components/session_websocket';
 import apiDispatcher from './dispatchers/api_dispatcher';
 import mainDispatcher from './dispatchers/main_dispatcher';
 import analyticsDispatcher from './dispatchers/analytics_dispatcher';
-import RetroReducer from './reducers/retro-reducer';
+import RetroReducer from './redux/reducers/retro-reducer';
 import * as ReduxActionDispatcher from './dispatchers/redux-action-dispatcher';
-import MessageReducer from './reducers/message_reducer';
-import ArchiveMiddleware from './reducers/archive-retro-middleware';
-import UserReducer from './reducers/user_reducer';
-import ConfigReducer from './reducers/config_reducer';
+import MessageReducer from './redux/reducers/message_reducer';
+import ArchiveMiddleware from './redux/middleware/archive-retro-middleware';
+import UserReducer from './redux/reducers/user_reducer';
+import ConfigReducer from './redux/reducers/config_reducer';
 
 const muiTheme = getMuiTheme({
   fontFamily: 'Karla',
