+++ web/src/redux/middleware/archive-retro-middleware.test.js
@@ -29,7 +29,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import '../spec_helper';
+import '../../spec_helper';
 import ArchiveMiddleware from './archive-retro-middleware';
 
 describe('ArchiveMiddleware', () => {
